move AsOwned impls closer to their types
